refactor: extract ResponseForm
