+++ web/src/components/NewResponseForm/NewResponseForm.tsx
@@ -7,24 +7,13 @@ import {
   PublicEvent,
 } from 'types/graphql'
 
-import {
-  Form,
-  TextField,
-  EmailField,
-  FormError,
-  Submit,
-  useForm,
-  NumberField,
-  SelectField,
-} from '@redwoodjs/forms'
+import { useForm } from '@redwoodjs/forms'
 import { useMutation } from '@redwoodjs/web'
 
-import FormField from 'src/components/FormField/FormField'
 import Typ from 'src/components/Typ/Typ'
-import { isEmail } from 'src/logic/validation'
-import { fieldAttrs, formErrorAttrs } from 'src/styles/classes'
 
 import dayjs from '../../apiLib/dayjs'
+import ResponseForm from '../ResponseForm/ResponseForm'
 
 interface Props {
   event: Pick<PublicEvent, 'id' | 'title' | 'responseConfig'>
@@ -70,16 +59,15 @@ const NewResponseForm = (props: Props) => {
     mode: 'onTouched',
     defaultValues: { name: '', email: '', headCount: 1, comment: '' },
   })
-  const { formState } = formMethods
 
-  const [sentToEmail, setSentToEmail] = useState<string | null>(null)
+  const [createdForEmail, setCreatedForEmail] = useState<string | null>(null)
 
   const [create, { loading, error }] = useMutation<
     CreateResponseMutation,
     CreateResponseMutationVariables
   >(CREATE_RESPONSE, {
     onCompleted: (data) => {
-      setSentToEmail(data.createResponse.email)
+      setCreatedForEmail(data.createResponse.email)
     },
   })
 
@@ -87,15 +75,14 @@ const NewResponseForm = (props: Props) => {
 
   const Title = <Typ x="head">RSVP to {event.title}</Typ>
 
-  if (sentToEmail) {
+  if (createdForEmail) {
     return (
       <>
         {Title}
-
         <Typ x="p">
           A confirmation email has been sent to you at{' '}
-          <strong>{sentToEmail}</strong>. Once you click the confirmation link
-          in your email, we will confirm your attendance and notify the
+          <strong>{createdForEmail}</strong>. Once you click the confirmation
+          link in your email, we will confirm your attendance and notify the
           organizer.
         </Typ>
         <Typ x="p">Thanks for using Freevite!</Typ>
@@ -103,109 +90,20 @@ const NewResponseForm = (props: Props) => {
     )
   }
 
-  const privacyNote = (() => {
-    switch (event.responseConfig) {
-      case 'SHOW_ALL':
-        return (
-          'Your name, head count, and comments will be shared ' +
-          'with the organizer and other guests.'
-        )
-      case 'SHOW_COUNTS_ONLY':
-        return 'Only your head count will be shared with the organizer and other guests.'
-      default:
-        return 'Your response will only be shared with the organizer.'
-    }
-  })()
-
-  // TODO: Form to request resend link(s) for email address
-  // TODO: Hold RSVP locally with cookie
-
   return (
     <>
       {Title}
-      <Typ x="p">{privacyNote}</Typ>
-      <Typ x="p">
-        Want to edit your RSVP? Use the confirmation link in your email.
-      </Typ>
-
-      <Form
-        className="mt-3"
+      <ResponseForm
+        event={event}
+        loading={loading}
+        error={error}
         formMethods={formMethods}
         onSubmit={(state: FormValues) =>
           create({
             variables: { eventId: event.id, input: stateToInput(state) },
           })
         }
-      >
-        <FormField name="email" text="Email Address*">
-          <Typ x="labelDetails">
-            We will send you a link to confirm. Your email will not be shared
-            with anyone else.
-          </Typ>
-          <EmailField
-            name="email"
-            validation={{ ...isEmail }}
-            disabled={loading}
-            {...fieldAttrs.input}
-          />
-        </FormField>
-
-        <FormField name="name" text="Your Name*">
-          <TextField
-            name="name"
-            validation={{ required: true }}
-            disabled={loading}
-            {...fieldAttrs.input}
-          />
-        </FormField>
-
-        <FormField name="headCount" text="Head Count*">
-          <Typ x="labelDetails">
-            How many people are you bringing, including yourself?
-          </Typ>
-          <NumberField
-            name="headCount"
-            disabled={loading}
-            validation={{
-              required: true,
-              min: { value: 1, message: 'Must be at least 1' },
-            }}
-            {...fieldAttrs.input}
-          />
-        </FormField>
-
-        <FormField name="comment" text="Additional Comments">
-          <Typ x="labelDetails">
-            Is there anything you want the organizer{' '}
-            {event.responseConfig === 'SHOW_ALL' && 'and other guests '}to know?
-          </Typ>
-          <TextField name="comment" disabled={loading} {...fieldAttrs.input} />
-        </FormField>
-
-        <FormField name="reminder" text="Send Me a Reminder">
-          <Typ x="labelDetails">
-            If you want, we can email you a reminder so you don&apos;t forget
-            about this event.
-          </Typ>
-          <SelectField name="reminder" disabled={loading} {...fieldAttrs.input}>
-            {Object.entries(reminderDurations).map(([label, value]) => (
-              <option key={label} value={label}>
-                {value
-                  ? `Remind me ${label} before the event starts`
-                  : 'Do not send me a reminder'}
-              </option>
-            ))}
-          </SelectField>
-        </FormField>
-
-        <Submit
-          className="button is-success mt-3"
-          disabled={loading || !formState.isValid}
-        >
-          {loading ? 'Submitting...' : 'Submit RSVP'}
-        </Submit>
-        <FormError error={error} {...formErrorAttrs} />
-      </Form>
+      />
     </>
   )
 }
